Acertos na parte de orçamento do perfil de localidade

+++ database/2.8.0/07_vw_perfil_localidade_orcamento_media_nacional.sql
@@ -10,7 +10,7 @@ FROM (
 		SUM(empenhado) AS empenhado,
 		localidade
 	FROM analysis.vw_perfil_localidade_orcamento
-	WHERE localidade::INTEGER BETWEEN 0 AND 9
+	WHERE localidade BETWEEN 0 AND 9
 	GROUP BY localidade
 ) AS a
 
@@ -25,7 +25,7 @@ FROM (
 		SUM(empenhado) AS empenhado,
 		localidade
 	FROM analysis.vw_perfil_localidade_orcamento
-	WHERE localidade::INTEGER BETWEEN 10 AND 99
+	WHERE localidade BETWEEN 10 AND 99
 	GROUP BY localidade
 ) AS a
 
@@ -40,6 +40,6 @@ FROM (
 		SUM(empenhado) AS empenhado,
 		localidade
 	FROM analysis.vw_perfil_localidade_orcamento
-	WHERE localidade::INTEGER > 99
+	WHERE localidade > 99
 	GROUP BY localidade
 ) AS a;
