+++ database/2.8.0/08_vw_perfil_localidade_orcamento_media_nacional_por_ano.sql
@@ -7,7 +7,7 @@ SELECT
 	(SUM(empenhado) / COUNT(*)) AS media,
 	COUNT(*) AS quantidade_localidades
 FROM analysis.vw_perfil_localidade_orcamento
-WHERE localidade::INTEGER BETWEEN 0 AND 9
+WHERE localidade BETWEEN 0 AND 9
 GROUP BY ano
 
 UNION
@@ -18,7 +18,7 @@ SELECT
 	(SUM(empenhado) / COUNT(*)) AS media,
 	COUNT(*) AS quantidade_localidades
 FROM analysis.vw_perfil_localidade_orcamento
-WHERE localidade::INTEGER BETWEEN 10 AND 99
+WHERE localidade BETWEEN 10 AND 99
 GROUP BY ano
 
 UNION
@@ -29,5 +29,5 @@ SELECT
 	(SUM(empenhado) / COUNT(*)) AS media,
 	COUNT(*) AS quantidade_localidades
 FROM analysis.vw_perfil_localidade_orcamento
-WHERE localidade::INTEGER > 99
+WHERE localidade > 99
 GROUP BY ano;
